backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
motor==3.3.1
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0
pymongo==4.6.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
httpx==0.25.0
bs4==0.0.1
newspaper3k==0.2.8
nltk==3.9
transformers==4.48.0
python-dotenv==1.0.0
openai==0.28.1
google-cloud-storage==2.10.0